refactor: revert generic injectIssueButton, consolidate CSS into plugin
Revert the over-abstracted injectIssueButton utility back to direct
observeQuerySelector calls in the plugin. Merge button positioning
CSS from utils/ into the plugin's index.module.css and remove the
now-unused utils/inject-issue-key-button.css.
