+++ plugins/copyIssueCacooFormat/index.module.css
@@ -1,4 +1,18 @@
-.cacooButton {
+/* 課題詳細ページ: #copyKey-help 横のボタン配置 */
+:global(.ticket__key) {
+	& > :global(.ticket__key-copy:not(:first-of-type)) {
+		transform: translateX(calc((100% + 2px) * var(--nth)));
+	}
+
+	& :nth-child(2 of :global(.ticket__key-copy)) { --nth: 1; }
+	& :nth-child(3 of :global(.ticket__key-copy)) { --nth: 2; }
+	& :nth-child(4 of :global(.ticket__key-copy)) { --nth: 3; }
+	& :nth-child(5 of :global(.ticket__key-copy)) { --nth: 4; }
+	& :nth-child(6 of :global(.ticket__key-copy)) { --nth: 5; }
+}
+
+/* 課題一覧ページ: cell-hover-actions 内のボタン */
+.button {
 	background: transparent;
 	color: var(--iconColorDefault);
 	border: 2px solid currentColor;
@@ -12,11 +26,11 @@
 	justify-content: center;
 }
 
-.cacooButton:hover {
+.button:hover {
 	color: var(--defaultColorAccent);
 }
 
-.cacooButton > svg {
+.button > svg {
 	width: 100%;
 	height: 100%;
 	object-fit: contain;
